Converting this repo to Avalon instance
Our setup (as of this writing) is to have a repository we own (not a fork) that
is our instance of Avalon, with our customizations. When we need to pull in upstream
changes, we will add the avalonmediasystem/avalon repository as a remote, and merge
changes from Github that way.