Remove whitespace for newlines in Subject facet
The CSS style white-space:pre-line; is causing newline characters to be
preserved as line breaks in the display. This is the quick and dirty fix, but
should be followed up with removing newline characters for the subject_sim
field on MediaObjects at time of ingest, along with a re-indexing job to fix
existing solr docs.

Fixes #155.